Revert "Merge branch 'develop' into 'master'"

See merge request s11-final/S11P31C108!66

+++ backend/kickboard/src/main/java/kb/user/api/response/LocationResponse.java
@@ -10,7 +10,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 @Schema(description = "지역 정보 응답")
 public record LocationResponse(
         @Schema(description = "지역 키")
-        int locationId,
+        Long locationId,
 
         @Schema(description = "지역명")
         String field
